Removed automatic, silent and erroneous database initialization
Previously, the ORM's automatic schema creation feature was incorrectly
used to bring up the database. This effectively broke applying
migrations so the whole README and Heroku setup have been updated
accordingly.

It is now recommended to use database migrations to bring up the
database:

    scripts/diamond.sh db upgrade

CLI commands remain so as to ease development and testing.